debugged and fixed issues with sequential file buffers. Rewind
Summary:
1. sequential file buffer management
2. rewind functionality
3. some logs for error cases.
4. making the time range of a blob log pick up correct option

Test Plan:

Reviewers:

Subscribers:

Tasks:

Blame Revision:

+++ utilities/blob_db/blob_db.h
@@ -46,7 +46,7 @@ struct BlobDBOptions {
   // first bucket is 1471542000 - 1471542600
   // second bucket is 1471542600 - 1471543200
   // and so on
-  uint64_t ttl_range;
+  uint32_t ttl_range;
 
   // at what size will the blobs be stored in separate log rather than
   // inline
